refactor service_manager and service_definition
